Rename test properly
