+++ apps/src/tests/index.ts
@@ -123,7 +123,7 @@ export { default as Test2631 } from './Test2631';
 export { default as Test2668 } from './Test2668';
 export { default as Test2675 } from './Test2675';
 export { default as Test2717 } from './Test2717';
-export { default as TestRepro } from './TestRepro';
+export { default as Test2767 } from './Test2767';
 export { default as TestScreenAnimation } from './TestScreenAnimation';
 export { default as TestScreenAnimationV5 } from './TestScreenAnimationV5';
 export { default as TestHeader } from './TestHeader';
